refactor: move banner data type to shared types (PIN-10357)
