+++ src/api/productUpdates/productUpdates.services.ts
@@ -1,11 +1,11 @@
 import { APP_MODE, INTEROP_RESOURCES_BASE_URL } from '@/config/env'
-import type { BannerData } from '@/hooks/bannerHooks/utils'
+import type { BannerData } from '@/types/banner.types'
 import axios from 'axios'
 
 async function getProductUpdatesJson() {
   // Mock data in development to avoid CORS issues with S3
   if (APP_MODE === 'development') {
-    return {
+    const bannerData: BannerData = {
       start: {
         date: '2023-05-25',
         time: '10:47',
@@ -14,7 +14,9 @@ async function getProductUpdatesJson() {
         date: '2028-05-25',
         time: '12:47',
       },
-    } as BannerData
+    }
+
+    return bannerData
   }
 
   const response = await axios.get<BannerData>(
