+++ src/types/banner.types.ts
@@ -1,0 +1,4 @@
+export type BannerData = {
+  start: { date: string; time: string }
+  end: { date: string; time: string }
+}
